printf: Cast format width to int

+++ lib/http_digest.c
@@ -149,7 +149,7 @@ CURLcode Curl_output_digest(struct connectdata *conn,
     tmp = strchr((char *)uripath, '?');
     if(tmp) {
       size_t urilen = tmp - (char *)uripath;
-      path = (unsigned char *) aprintf("%.*s", urilen, uripath);
+      path = (unsigned char *) aprintf("%.*s", (int)urilen, uripath);
     }
   }
   if(!tmp)
