+++ src/tool_getparam.c
@@ -1411,7 +1411,7 @@ ParameterError getparameter(const char *flag, /* f or -long-flag */
               return PARAM_NO_MEM;
             }
             if(nlen > 0) { /* only append '=' if we have a name */
-              msnprintf(n, outlen, "%.*s=%s", nlen, nextarg, enc);
+              msnprintf(n, outlen, "%.*s=%s", (int)nlen, nextarg, enc);
               size = outlen-1;
             }
             else {
